fix login

+++ routes/bugs.js
@@ -68,16 +68,16 @@ router.put('/:id', (req, res) => {
 // @route   DELETE api/bugs/:id
 // @desc    Delete Bugs
 // @access  Private
-router.delete('/:id', async (req, res) => {
+router.delete('/:id', auth,async (req, res) => {
   try {
     let bug = await Bug.findById(req.params.id);
 
     if (!bug) return res.status(404).json({ msg: 'Contact not found' });
-
+    
     // Make sure user owns contact
-    // if (contact.user.toString() !== req.user.id) {
-    //   return res.status(401).json({ msg: 'Not authorized' });
-    // }
+    if (bug.user.toString() !== req.user.id) {
+      return res.status(401).json({ msg: 'Not authorized' });
+    }
 
     await Bug.findByIdAndRemove(req.params.id);
